Revert "verification and sign in ui"
