+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { Route, Routes } from "react-router-dom";
-import SignUp from "./routes/authentication/Authentication.component";
+import SignUp from "./routes/sign-up/SignUp.component";
 import Navbar from "./shared/ components/Navbar";
 
 const App = () => {
@@ -7,9 +7,7 @@ const App = () => {
     <>
       <Navbar />
       <Routes>
-        {["/sign-in", "/sign-up"].map((path, index) => (
-          <Route path={path} element={<SignUp />} key={index} />
-        ))}
+        <Route path="/sign-up" element={<SignUp />} />
       </Routes>
     </>
   );
