--- src/shared/ components/Container.tsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  padding: var(--spacing-md);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
